# src/docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Evently API",
    "description": "API for Evently",
    "version": "1.0.0"
  },
  "host": [
    "theodev.myftp.org:89"
  ],
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your Bearer token in the format **Bearer <token>**"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/categories": {
      "post": {
        "tags": [
          "Category"
        ],
        "description": "Create a new category",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Category details",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Example"
                },
                "icon": {
                  "type": "string",
                  "example": "icon"
                },
                "color": {
                  "type": "string",
                  "example": "#FFFFFF"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Category"
            },
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "get": {
        "tags": [
          "Category"
        ],
        "description": "Get user categories",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Category"
            },
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/categories/default": {
      "get": {
        "tags": [
          "Category"
        ],
        "description": "Get default categories",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/Category"
            },
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/categories/{id}": {
      "put": {
        "tags": [
          "Category"
        ],
        "description": "Update category",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Updated name"
                },
                "icon": {
                  "type": "string",
                  "example": "updated-icon"
                },
                "color": {
                  "type": "string",
                  "example": "#000000"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      },
      "delete": {
        "tags": [
          "Category"
        ],
        "description": "Delete category",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Category": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 666
        },
        "name": {
          "type": "string",
          "example": "foobar"
        },
        "icon": {
          "type": "string",
          "example": "foobar"
        },
        "color": {
          "type": "string",
          "example": "foobar"
        },
        "default": {
          "type": "boolean",
          "example": true
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      }
    },
...
    "Event": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 666
        },
        "name": {
          "type": "string",
          "example": "foobar"
        },
        "description": {
          "type": "string",
          "example": "foobar"
        },
        "location": {
          "type": "string",
          "example": "foobar"
        },
        "childCreated": {
          "type": "boolean",
          "example": true
        },
        "recurrencePattern": {
          "type": "string",
          "example": "monthly"
        },
        "interval": {
          "type": "number",
          "example": 666
        },
        "maxOccurence": {
          "type": "number",
          "example": 666
        },
        "targetDate": {
          "type": "string",
          "example": "2019-01-01T00:00:00.000Z"
        },
        "inviteToken": {
          "type": "string",
          "example": "foobar"
        },
        "categoryID": {
          "type": "number",
          "example": 666
        },
        "userID": {
          "type": "number",
          "example": 666
        },
        "folderID": {
          "type": "number",
          "example": 666
        },
        "createdAt": {
          "type": "string",
          "example": "2019-01-01T00:00:00.000Z"
        },
        "deletedAt": {
          "type": "string",
          "example": "2019-01-01T00:00:00.000Z"
        },
        "joinedUser": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/JoinedEventEntity"
          }
        }
      }
    },
...
    "JoinedEventEntity": {
      "type": "object",
      "properties": {
        "leaveDate": {
          "type": "string",
          "example": "2019-01-01T00:00:00.000Z"
        },
        "invitationStatus": {
          "type": "string",
          "example": "invited"
        },
        "event": {
          "$ref": "#/definitions/Event"
        },
        "user": {
          "$ref": "#/definitions/User"
        }
      }
    },
...
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number",
          "example": 666
        },
        "username": {
          "type": "string",
          "example": "foobar"
        },
        "firstName": {
          "type": "string",
          "example": "foobar"
        },
        "lastName": {
          "type": "string",
          "example": "foobar"
        },
        "email": {
          "type": "string",
          "example": "foobar"
        },
        "sub": {
          "type": "string",
          "example": "foobar"
        },
        "password": {
          "type": "string",
          "example": "foobar"
        },
        "provider": {
          "type": "string",
          "example": "foobar"
        },
        "isDeleted": {
          "type": "boolean",
          "example": true
        },
        "isCompleted": {
          "type": "boolean",
          "example": true
        },
        "createdAt": {
          "type": "string",
          "example": "2019-01-01T00:00:00.000Z"
        },
        "deletedAt": {
          "type": "string",
          "example": "2019-01-01T00:00:00.000Z"
        },
        "categories": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Category"
          }
        },
        "events": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Event"
          }
        },
        "joinedEvents": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/JoinedEventEntity"
          }
        }
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}